Admin col updates for consistency
- remove cols that got duped up in client tab
- use badge renderer consistently for these new id-ish things
- re-order touched col files to unscramble, make it easier to scan/find cols
