+++ admin/columns/Core.ts
@@ -11,15 +11,29 @@ import {dateTimeRenderer} from '@xh/hoist/format';
 import {Icon} from '@xh/hoist/icon';
 import copy from 'clipboard-copy';
 
-export const name: ColumnSpec = {
-    field: {name: 'name', type: 'string'},
-    width: 200
+export const badgeCol: ColumnSpec = {
+    autosizable: false,
+    width: 100,
+    renderer: badgeRenderer
 };
 
-export const type: ColumnSpec = {
-    field: {name: 'type', type: 'string'},
-    width: 100
-};
+export function badgeRenderer(v) {
+    return v
+        ? badge({
+              item: v,
+              className: 'xh-badge-col',
+              style: {cursor: 'copy'},
+              title: 'Double-click to copy',
+              onDoubleClick: () => {
+                  copy(v);
+                  XH.toast({
+                      icon: Icon.copy(),
+                      message: `Copied ${v}`
+                  });
+              }
+          })
+        : '-';
+}
 
 export const description: ColumnSpec = {
     field: {name: 'description', type: 'string'},
@@ -27,11 +41,9 @@ export const description: ColumnSpec = {
     minWidth: 200
 };
 
-export const notes: ColumnSpec = {
-    field: {name: 'notes', type: 'string'},
-    minWidth: 60,
-    flex: true,
-    tooltip: true
+export const name: ColumnSpec = {
+    field: {name: 'name', type: 'string'},
+    width: 200
 };
 
 export const note: ColumnSpec = {
@@ -45,33 +57,20 @@ export const note: ColumnSpec = {
     tooltip: true
 };
 
+export const notes: ColumnSpec = {
+    field: {name: 'notes', type: 'string'},
+    minWidth: 60,
+    flex: true,
+    tooltip: true
+};
+
 export const timestampNoYear: ColumnSpec = {
     field: {name: 'timestamp', type: 'date'},
     ...dateTimeSec,
     renderer: dateTimeRenderer({fmt: 'MMM DD HH:mm:ss.SSS'})
 };
 
-export function badgeRenderer(v) {
-    return v
-        ? badge({
-              item: v,
-              className: 'xh-font-family-mono',
-              style: {cursor: 'copy'},
-              intent: 'primary',
-              title: 'Double-click to copy',
-              onDoubleClick: () => {
-                  copy(v);
-                  XH.toast({
-                      icon: Icon.copy(),
-                      message: `Copied ${v}`
-                  });
-              }
-          })
-        : '-';
-}
-
-export const badgeCol: ColumnSpec = {
-    autosizable: false,
-    width: 90,
-    renderer: badgeRenderer
+export const type: ColumnSpec = {
+    field: {name: 'type', type: 'string'},
+    width: 100
 };
