+++ admin/tabs/activity/tracking/detail/ActivityDetailView.ts
@@ -4,11 +4,11 @@
  *
  * Copyright © 2025 Extremely Heavy Industries Inc.
  */
-import {correlationId, instance} from '@xh/hoist/admin/columns';
+import {badgeRenderer} from '@xh/hoist/admin/columns';
 import {form} from '@xh/hoist/cmp/form';
 import {grid, gridCountLabel} from '@xh/hoist/cmp/grid';
 import {a, div, filler, h3, hframe, placeholder, span} from '@xh/hoist/cmp/layout';
-import {hoistCmp, creates} from '@xh/hoist/core';
+import {creates, hoistCmp} from '@xh/hoist/core';
 import {colChooserButton, exportButton} from '@xh/hoist/desktop/cmp/button';
 import {formField} from '@xh/hoist/desktop/cmp/form';
 import {gridFindField} from '@xh/hoist/desktop/cmp/grid';
@@ -96,10 +96,12 @@ const detailRecForm = hoistCmp.factory<ActivityDetailModel>(({model}) => {
                               }
                           }),
                           formField({
-                              field: 'loadId'
+                              field: 'loadId',
+                              readonlyRenderer: badgeRenderer
                           }),
                           formField({
-                              field: 'tabId'
+                              field: 'tabId',
+                              readonlyRenderer: badgeRenderer
                           }),
                           formField({
                               field: 'url',
@@ -107,11 +109,11 @@ const detailRecForm = hoistCmp.factory<ActivityDetailModel>(({model}) => {
                           }),
                           formField({
                               field: 'instance',
-                              readonlyRenderer: v => instance.renderer(v, null)
+                              readonlyRenderer: badgeRenderer
                           }),
                           formField({
                               field: 'correlationId',
-                              readonlyRenderer: v => correlationId.renderer(v, null)
+                              readonlyRenderer: badgeRenderer
                           }),
                           formField({
                               field: 'elapsed',
